Move caching tests to their own file
