+++ test/test_arhp.rb
@@ -74,33 +74,6 @@ class ActiveRecordHostPoolTest < Minitest::Test
     assert_equal(Pool1DbA.connection.raw_connection, Pool1DbC.connection.raw_connection)
   end
 
-  if ActiveRecord.version >= Gem::Version.new('6.0')
-    def test_models_with_matching_hosts_and_non_matching_databases_issue_exists_without_arhp_patch
-      simulate_rails_app_active_record_railties
-
-      # Remove patch that fixes an issue in Rails 6+ to ensure it still
-      # exists. If this begins to fail then it may mean that Rails has fixed
-      # the issue so that it no longer occurs.
-      without_module_patch(ActiveRecordHostPool::ClearQueryCachePatch, :clear_query_caches_for_current_thread) do
-        exception = assert_raises(ActiveRecord::StatementInvalid) do
-          ActiveRecord::Base.cache { Pool1DbC.create! }
-        end
-
-        assert_equal("Mysql2::Error: Table 'arhp_test_db_b.pool1_db_cs' doesn't exist", exception.message)
-      end
-    end
-
-    def test_models_with_matching_hosts_and_non_matching_databases_do_not_mix_up_underlying_database
-      simulate_rails_app_active_record_railties
-
-      # ActiveRecord 6.0 introduced a change that surfaced a problematic code
-      # path in active_record_host_pool when clearing caches across connection
-      # handlers which can cause the database to change.
-      # See ActiveRecordHostPool::ClearQueryCachePatch
-      ActiveRecord::Base.cache { Pool1DbC.create! }
-    end
-  end
-
   def test_connection_returns_a_proxy
     assert_kind_of ActiveRecordHostPool::ConnectionProxy, Pool1DbA.connection
   end
@@ -118,14 +91,6 @@ class ActiveRecordHostPoolTest < Minitest::Test
     refute Pool1DbA.connection.respond_to?(:test_private_method)
     assert_includes(Pool1DbA.connection.private_methods, :test_private_method)
     assert_equal true, Pool1DbA.connection.send(:test_private_method)
-  end
-
-  def test_should_not_share_a_query_cache
-    Pool1DbA.create(val: 'foo')
-    Pool1DbB.create(val: 'foobar')
-    Pool1DbA.connection.cache do
-      refute_equal Pool1DbA.first.val, Pool1DbB.first.val
-    end
   end
 
   def test_object_creation
